fix cpp-vector.md article vector link
Linked vector class is no longer experimental on 'master' branch,
for this reason link to the vector in this article returns 404.
cpp-vector.md article says that the vector was added recently.
The link to libpmemobj-cpp vector class was changed from 'master'
to the earliest available version of the lib with vector marked
as experimental.